Initial scalability experiments
Create a new project with a barebones project structure using `uv`
to manage dependencies.

Add a script, quick_test.py, to run a simple scalability test of
an approach using Python asyncio. The results show that a naive
single-core asyncio solution can get to over 50e3 messages/second.
That's enough to message the entire US population in under two hours,
or the population of Massachusetts in under 3 minutes.